feature/US20909 Fix Site Title
For some articles that contained meta title information in their
markdown front matter, the site title was being rendered as only the
name of the article. Insights team requested that all articles have
"| Crossroads Media" appended to the end of the title. This required a
change in the `jekyll-crds` github library that the crdschurch
organization manages.

This commit bumps the `jekyll-crds` dependency up one bugfix version to
2.3.2