HOTFIX US22151 Fix Location Finder Endpoint
This commit changes the logic and hardcodes the endpoint from which
location data is retrieved. There was consideration toward using a new
env variable, but since the old Gateway API endpoint is getting revamped
one piece at a time, it was decided to hard code until the changeover is
complete, reducing the number of env vars that get created.